+++ src/matterbridgeEndpoint.ts
@@ -160,6 +160,7 @@ import {
   getDefaultPowerSourceRechargeableBatteryClusterServer,
   getDefaultDeviceEnergyManagementClusterServer,
   getDefaultDeviceEnergyManagementModeClusterServer,
+  getDefaultPowerSourceBatteryClusterServer,
 } from './matterbridgeEndpointHelpers.js';
 
 export class MatterbridgeEndpoint extends Endpoint {
@@ -894,6 +895,30 @@ export class MatterbridgeEndpoint extends Endpoint {
   }
 
   /**
+   * Creates a default power source battery cluster server.
+   *
+   * @param {null | number} batPercentRemaining - The remaining battery percentage (default: null). The attribute is in the range 0-200.
+   * @param {PowerSource.BatChargeLevel} batChargeLevel - The battery charge level (default: PowerSource.BatChargeLevel.Ok).
+   * @param {null | number} batVoltage - The battery voltage (default: null).
+   * @param {PowerSource.BatReplaceability} batReplaceability - The replaceability of the battery (default: PowerSource.BatReplaceability.Unspecified).
+   * @returns {this} The current MatterbridgeEndpoint instance for chaining.
+   *
+   * @remarks
+   * - order: The order of the power source is a persisted attribute that indicates the order in which the power sources are used.
+   * - description: The description of the power source is a fixed attribute that describes the power source type.
+   * - batReplaceability: The replaceability of the battery is a fixed attribute that indicates whether the battery is user-replaceable or not.
+   */
+  createDefaultPowerSourceBatteryClusterServer(
+    batPercentRemaining: null | number = null,
+    batChargeLevel: PowerSource.BatChargeLevel = PowerSource.BatChargeLevel.Ok,
+    batVoltage: null | number = null,
+    batReplaceability: PowerSource.BatReplaceability = PowerSource.BatReplaceability.Unspecified,
+  ): this {
+    this.behaviors.require(MatterbridgePowerSourceServer.with(PowerSource.Feature.Battery), getDefaultPowerSourceBatteryClusterServer(batPercentRemaining, batChargeLevel, batVoltage, batReplaceability));
+    return this;
+  }
+
+  /**
    * Creates a default power source replaceable battery cluster server.
    *
    * @param {number} batPercentRemaining - The remaining battery percentage (default: 100). The attribute is in the range 0-200.
@@ -901,7 +926,7 @@ export class MatterbridgeEndpoint extends Endpoint {
    * @param {number} batVoltage - The battery voltage (default: 1500).
    * @param {string} batReplacementDescription - The description of the battery replacement (default: 'Battery type').
    * @param {number} batQuantity - The quantity of the battery (default: 1).
-   * @param {PowerSource.BatReplaceability} batReplaceability - The replaceability of the battery (default: PowerSource.BatReplaceability.Unspecified).
+   * @param {PowerSource.BatReplaceability} batReplaceability - The replaceability of the battery (default: PowerSource.BatReplaceability.UserReplaceable).
    * @returns {this} The current MatterbridgeEndpoint instance for chaining.
    *
    * @remarks